refactor(router): start use nested routes from react-router v6
