+++ src/app/routing/routes/index.ts
@@ -1,0 +1,5 @@
+export * from './profile-admin';
+export * from './profile-recipient';
+export * from './profile-volonteer';
+export * from './public-profile';
+export * from './public-routes';
